web/requirements.txt
flask>=3.0
gunicorn>=21.2
python-telegram-bot>=21.0
requests>=2.31
